+++ src/sciform/formatter.py
@@ -89,8 +89,11 @@ class Formatter:
       exponent mode. When ``True``, the float is multipled by 100 and
       a % symbol is appended to the end of the string after formatting.
     :param superscript_exp: ``bool`` indicating if the exponent string
-      should be converted into superscript notation. E.g. `1.23e+02` is
-      converted to `1.23×10²`
+      should be converted into superscript notation. E.g. ``'1.23e+02'``
+      is converted to ``'1.23×10²'``
+    :param latex: ``bool`` indicating if the resulting string should be
+      converted into a latex parseable code, e.g.
+      ``'\\left(1.23 \\pm 0.01\\right)\\times 10^{2}'``.
     :param nan_inf_exp: ``bool`` indicating whether non-finite floats
       such as ``float('nan')`` or ``float('inf')`` should be formatted
       with exponent symbols when exponent modes including exponent
@@ -137,6 +140,7 @@ class Formatter:
             capitalize: bool = None,
             percent: bool = None,
             superscript_exp: bool = None,
+            latex: bool = None,
             nan_inf_exp: bool = None,
             use_prefix: bool = None,
             extra_si_prefixes: dict[int, str] = None,
@@ -164,6 +168,7 @@ class Formatter:
             capitalize=capitalize,
             percent=percent,
             superscript_exp=superscript_exp,
+            latex=latex,
             nan_inf_exp=nan_inf_exp,
             use_prefix=use_prefix,
             extra_si_prefixes=extra_si_prefixes,
@@ -201,6 +206,7 @@ class Formatter:
                    capitalize=options.capitalize,
                    percent=options.percent,
                    superscript_exp=options.superscript_exp,
+                   latex=options.latex,
                    nan_inf_exp=options.nan_inf_exp,
                    use_prefix=options.use_prefix,
                    extra_si_prefixes=options.extra_si_prefixes,